requirements.txt
Pillow==7.1.2
asgiref==3.2.10
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
django-admin-thumbnails==0.2.5
django-ckeditor==5.9.0
django-currencies==0.10.1
django-js-asset==1.2.2
django-mptt==0.11.0
idna==2.9
jsonfield==2.1.1
python-gettext==4.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
